Convert restrictToColors from single-color to bitset for AttackFreeCell, ReconnectFreeCell, ReconnectCreature, and Receiver

+++ source/EngineGpuKernels/Entities.cuh
@@ -174,7 +174,7 @@ struct Generator
 
 struct AttackFreeCell
 {
-    uint8_t restrictToColors;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColors;  // bitset: bit i set = color i allowed, 0x3FF = all colors
 };
 
 struct AttackCreature
@@ -292,14 +292,14 @@ struct ReconnectStructure
 
 struct ReconnectFreeCell
 {
-    uint8_t restrictToColors;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColors;  // bitset: bit i set = color i allowed, 0x3FF = all colors
 };
 
 struct ReconnectCreature
 {
     uint32_t minNumCells;     // 0 = no restriction
     uint32_t maxNumCells;     // 0 = no restriction
-    uint8_t restrictToColors;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColors;  // bitset: bit i set = color i allowed, 0x3FF = all colors
     LineageRestriction restrictToLineage;
 };
 
@@ -383,7 +383,7 @@ struct Sender
 
 struct Receiver
 {
-    uint8_t restrictToColors;  // 0 ... 6 = color restriction, 255 = no restriction
+    uint16_t restrictToColors;  // bitset: bit i set = color i allowed, 0x3FF = all colors
     LineageRestriction restrictToLineage;
 };
